+++ src/php/DataSift/Storyplayer/Cli/Feature/UbLangConsoleSwitch.php
@@ -43,21 +43,12 @@
 
 namespace DataSift\Storyplayer\Cli;
 
-use Exception;
-use stdClass;
 use Phix_Project\CliEngine;
-use Phix_Project\CliEngine\CliCommand;
-use Phix_Project\ExceptionsLib1\Legacy_ErrorHandler;
-use Phix_Project\ExceptionsLib1\Legacy_ErrorException;
-use DataSift\Stone\ConfigLib\E5xx_ConfigFileNotFound;
-use DataSift\Stone\ConfigLib\E5xx_InvalidConfigFile;
-use DataSift\Storyplayer\PlayerLib\E4xx_NoSuchReport;
-use DataSift\Storyplayer\PlayerLib\StoryTeller;
-use DataSift\Storyplayer\Console\DevModeConsole;
-use DataSift\Storyplayer\Injectables;
+use Phix_Project\CliEngine\CliResult;
+use Phix_Project\CliEngine\CliSwitch;
 
 /**
- * Support for the console
+ * Tell Storyplayer to switch to the 'Ubiquitous Language' console
  *
  * @category  Libraries
  * @package   Storyplayer/Cli
@@ -66,54 +57,40 @@ use DataSift\Storyplayer\Injectables;
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://datasift.github.io/storyplayer
  */
-class Feature_ConsoleSupport implements Feature
+class Feature_UbLangConsoleSwitch extends CliSwitch
 {
-    public function addSwitches(CliCommand $command, $additionalContext)
+    public function __construct()
     {
-        $command->addSwitches([
-            new Feature_DevModeSwitch,
-            new Feature_UbLangConsoleSwitch,
-        ]);
+        // define our name, and our description
+        $this->setName('ublang');
+        $this->setShortDescription('use the Ubiquitous Language console');
+        $this->setLongDesc(
+            "This switch changes Storyplayer's default output to the Ubiquitous Language console."
+            . " The UbLang console provides a more descriptive description of what each of your tests is doing."
+            . PHP_EOL . PHP_EOL
+            . "The full story log is also available in storyplayer.log."
+        );
+
+        // what are the long switches?
+        $this->addLongSwitch('ublang');
+
+        // all done
     }
 
-    public function initBeforeModulesAvailable(CliEngine $engine, CliCommand $command, Injectables $injectables)
+    /**
+     *
+     * @param  CliEngine $engine
+     * @param  integer   $invokes
+     * @param  array     $params
+     * @param  boolean   $isDefaultParam
+     * @return CliResult
+     */
+    public function process(CliEngine $engine, $invokes = 1, $params = array(), $isDefaultParam = false)
     {
-        // shorthand
-        $output = $injectables->output;
+        // remember the setting
+        $engine->options->console = "UbLangConsole";
 
-        // we also want to create storyplayer.log every time
-        $devModeConsole = new DevModeConsole();
-        $devModeConsole->addOutputToFile('storyplayer.log');
-
-        // switch output plugins first, before we do anything else at all
-        if (isset($engine->options->dev) && $engine->options->dev) {
-            // switch our main output to 'dev mode'
-            $output->usePluginAsConsole($devModeConsole);
-            $devModeConsole->addOutputToStdout();
-
-            // all done
-            return;
-        }
-
-        // if we get here, then we only want the devMode console for our logfile
-        $output->usePluginInSlot($devModeConsole, 'logfile');
-
-        // now, do we want any other console?
-        if (isset($engine->options->console)) {
-            $consoleClass = "DataSift\\Storyplayer\\Console\\" . $engine->options->console;
-            if (!class_exists($consoleClass)) {
-                $output->logCliError("unable to find class '{$consoleClass}' to be our console");
-                exit(1);
-            }
-
-            $console = new $consoleClass;
-            $output->usePluginAsConsole($console);
-            $console->addOutputToStdout();
-        }
-    }
-
-    public function initAfterModulesAvailable(StoryTeller $st, CliEngine $engine, Injectables $injectables)
-    {
-        // no-op
+        // tell the engine that it is done
+        return new CliResult(CliResult::PROCESS_CONTINUE);
     }
 }